Caremada-Recommender-Engine-master/requirements.txt
PyInstaller
PyQt5
PyQt5-sip
altgraph
cycler
dnspython
fbs
future
joblib
kiwisolver
macholib
matplotlib
mysql-connector-python
numpy
pandas
pefile
pip
pymongo
pyparsing
python-dateutil
pytz
pywin32-ctypes
scikit-learn
scipy
setuptools
six
sklearn